refactor(sync): use PushType enum throughout incremental sync chain
- Update sync_cipher_by_id, sync_folder_by_id, sync_send_by_id to accept PushType
- Update execute_cipher_incremental, execute_folder_incremental, execute_send_incremental
- Update apply_*_incremental_update methods to use PushType for delete detection
- Replace hardcoded event type numbers (2, 8, 14) with enum variants
- Remove unnecessary i32 conversions in realtime_sync_service
- Update all log messages to use {:?} for PushType display

This completes the type-safe refactoring of WebSocket event handling,
eliminating all magic numbers from the incremental sync flow.

+++ src-tauri/src/application/services/realtime_sync_service.rs
@@ -557,7 +557,7 @@ impl RealtimeSyncService {
 
         match self
             .sync_service
-            .sync_cipher_by_id(command, String::from(cipher_id), push_type as i32)
+            .sync_cipher_by_id(command, String::from(cipher_id), push_type)
             .await
         {
             Ok(_) => EventOutcome::Continue,
@@ -611,7 +611,7 @@ impl RealtimeSyncService {
 
         match self
             .sync_service
-            .sync_folder_by_id(command, String::from(folder_id), push_type as i32)
+            .sync_folder_by_id(command, String::from(folder_id), push_type)
             .await
         {
             Ok(_) => EventOutcome::Continue,
@@ -665,7 +665,7 @@ impl RealtimeSyncService {
 
         match self
             .sync_service
-            .sync_send_by_id(command, String::from(send_id), push_type as i32)
+            .sync_send_by_id(command, String::from(send_id), push_type)
             .await
         {
             Ok(_) => EventOutcome::Continue,
